refactor: move one_hot_encoding/ to core/data/encoding.py
