+++ simba/core/data/encoding.py
